+++ R/glue.R
@@ -237,15 +237,6 @@ glue_collapse <- function(x, sep = "", width = Inf, last = "") {
   as_glue(x)
 }
 
-# nocov start
-#' @rdname glue-deprecated
-#' @export
-collapse <- function(x, sep = "", width = Inf, last = "") {
-  .Deprecated("glue_collapse", package = "glue")
-  glue_collapse(x, sep, width, last)
-}
-# nocov end
-
 #' Trim a character vector
 #'
 #' This trims a character vector according to the trimming rules used by glue.